refactor to organize code
