+++ hamt/hamt.js
