+++ hamt/hash.js
